chore(684): Move file and update imports
