+++ pages/mes-jeunes/[jeune_id]/actions/[action_id]/index.tsx
@@ -6,9 +6,9 @@ import React, { useState } from 'react'
 import InfoAction from 'components/action/InfoAction'
 import { RadioButtonStatus } from 'components/action/RadioButtonStatus'
 import FailureMessage from 'components/FailureMessage'
-import SuccessMessage from 'components/SuccessMessage'
 import Button, { ButtonStyle } from 'components/ui/Button'
 import IconComponent, { IconName } from 'components/ui/IconComponent'
+import SuccessMessage from 'components/ui/SuccessMessage'
 import { Action, StatutAction } from 'interfaces/action'
 import { UserStructure, UserType } from 'interfaces/conseiller'
 import { Jeune } from 'interfaces/jeune'
