+++ pages/mes-rendezvous/index.tsx
@@ -4,8 +4,8 @@ import { useRouter } from 'next/router'
 import React, { useState } from 'react'
 
 import RdvList from 'components/rdv/RdvList'
-import SuccessMessage from 'components/SuccessMessage'
 import ButtonLink from 'components/ui/ButtonLink'
+import SuccessMessage from 'components/ui/SuccessMessage'
 import Tab from 'components/ui/Tab'
 import TabList from 'components/ui/TabList'
 import { UserStructure } from 'interfaces/conseiller'
